agents/oracle/requirements.txt
functions-framework==3.*
Flask==3.0.3
gunicorn==21.2.0
google-cloud-storage==2.18.2
google-api-python-client==2.145.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
PyMuPDF==1.24.9
sendgrid==6.11.0
